Typescript updates
