+++ src/authprovider.js
@@ -1,4 +1,4 @@
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
+
 
 /**
  * Authentication Provider details.
