+++ src/browser.js
@@ -1,4 +1,4 @@
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
+
 
 /**
  * Platform dependant utilities for the openEO JS Client.
