--- src/builder/tapdigit.d.ts
@@ -1,20 +1,0 @@
-export function Lexer(): {
-    reset: (str: any) => void;
-    next: () => {
-        type: any;
-        value: any;
-        start: number;
-        end: number;
-    };
-    peek: () => {
-        type: any;
-        value: any;
-        start: number;
-        end: number;
-    };
-};
-export function Parser(): {
-    parse: (expression: any) => {
-        Expression: any;
-    };
-};
