--- src/env.d.ts
@@ -1,2 +1,0 @@
-declare const _exports: typeof import("./node") | typeof import("./browser");
-export = _exports;
