+++ src/job.js
@@ -2,8 +2,8 @@ const Environment = require('./env');
 const BaseEntity = require('./baseentity');
 const Logs = require('./logs');
 const Utils = require('@openeo/js-commons/src/utils');
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
-const { Link, Process } = require('./typedefs'); // eslint-disable-line no-unused-vars
+
+
 
 const STOP_STATUS = ['finished', 'canceled', 'error'];
