+++ src/oidcprovider.js
@@ -1,8 +1,8 @@
 const Utils = require('@openeo/js-commons/src/utils');
 const AuthProvider = require('./authprovider');
 const Oidc = require('oidc-client');
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
-const { Link } = require('./typedefs'); // eslint-disable-line no-unused-vars
+
+
 
 /**
  * The Authentication Provider for OpenID Connect.
