+++ src/service.js
@@ -1,7 +1,7 @@
 const BaseEntity = require('./baseentity');
 const Logs = require('./logs');
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
-const { Process } = require('./typedefs'); // eslint-disable-line no-unused-vars
+
+
 
 /**
  * A Secondary Web Service.
