+++ src/userfile.js
@@ -1,8 +1,8 @@
 const Environment = require('./env');
 const BaseEntity = require('./baseentity');
-const Connection = require('./connection'); // eslint-disable-line no-unused-vars
 
-const Stream = require('stream'); // eslint-disable-line no-unused-vars
+
+
 
 /**
  * A File on the user workspace.
